+++ java/tools/src/test/java/org/apache/arrow/tools/EchoServerTest.java
@@ -112,7 +112,6 @@ public class EchoServerTest {
         writer.writeBatch();
       }
       writer.end();
-      assertTrue(writer.getRecordBlocks().isEmpty());
 
       assertEquals(new Schema(asList(field)), reader.getVectorSchemaRoot().getSchema());
