+++ java/vector/src/main/java/org/apache/arrow/vector/ipc/ArrowFileWriter.java
@@ -20,19 +20,26 @@ package org.apache.arrow.vector.ipc;
 
 import java.io.IOException;
 import java.nio.channels.WritableByteChannel;
+import java.util.ArrayList;
 import java.util.List;
+
+import com.google.common.annotations.VisibleForTesting;
 
 import org.apache.arrow.vector.VectorSchemaRoot;
 import org.apache.arrow.vector.dictionary.DictionaryProvider;
 import org.apache.arrow.vector.ipc.message.ArrowBlock;
+import org.apache.arrow.vector.ipc.message.ArrowDictionaryBatch;
 import org.apache.arrow.vector.ipc.message.ArrowFooter;
-import org.apache.arrow.vector.types.pojo.Schema;
+import org.apache.arrow.vector.ipc.message.ArrowRecordBatch;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 public class ArrowFileWriter extends ArrowWriter {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(ArrowFileWriter.class);
+
+  private final List<ArrowBlock> dictionaryBlocks = new ArrayList<>();
+  private final List<ArrowBlock> recordBlocks = new ArrayList<>();
 
   public ArrowFileWriter(VectorSchemaRoot root, DictionaryProvider provider, WritableByteChannel out) {
     super(root, provider, out);
@@ -44,12 +51,23 @@ public class ArrowFileWriter extends ArrowWriter {
   }
 
   @Override
-  protected void endInternal(WriteChannel out,
-                             Schema schema,
-                             List<ArrowBlock> dictionaries,
-                             List<ArrowBlock> records) throws IOException {
+  protected ArrowBlock writeDictionaryBatch(ArrowDictionaryBatch batch) throws IOException {
+    ArrowBlock block = super.writeDictionaryBatch(batch);
+    dictionaryBlocks.add(block);
+    return block;
+  }
+
+  @Override
+  protected ArrowBlock writeRecordBatch(ArrowRecordBatch batch) throws IOException {
+    ArrowBlock block = super.writeRecordBatch(batch);
+    recordBlocks.add(block);
+    return block;
+  }
+
+  @Override
+  protected void endInternal(WriteChannel out) throws IOException {
     long footerStart = out.getCurrentPosition();
-    out.write(new ArrowFooter(schema, dictionaries, records), false);
+    out.write(new ArrowFooter(schema, dictionaryBlocks, recordBlocks), false);
     int footerLength = (int) (out.getCurrentPosition() - footerStart);
     if (footerLength <= 0) {
       throw new InvalidArrowFileException("invalid footer");
@@ -58,5 +76,15 @@ public class ArrowFileWriter extends ArrowWriter {
     LOGGER.debug(String.format("Footer starts at %d, length: %d", footerStart, footerLength));
     ArrowMagic.writeMagic(out, false);
     LOGGER.debug(String.format("magic written, now at %d", out.getCurrentPosition()));
+  }
+
+  @VisibleForTesting
+  public List<ArrowBlock> getRecordBlocks() {
+    return recordBlocks;
+  }
+
+  @VisibleForTesting
+  public List<ArrowBlock> getDictionaryBlocks() {
+    return dictionaryBlocks;
   }
 }
